+++ ChangeLog.md
@@ -59,6 +59,8 @@
 
   * Add support for dtype in `pl.to_json` and `pl.from_json` (Tim Bretl).
 
+  * Add better grading-instance autoscaling calculations (Matt West).
+
   * Fix broken file upload element (Nathan Walters).
 
   * Fix broken popover and improve assessment label styles (Nathan Walters).